Remove documentation
The documentation is written in the `README.md` too, so instead of
updating it both in the `README.md` and in `check-outdated.sh`, I
prefer update it in the `README.md` only.

The documentation is still available because in the header there
is the link to my github project which shown, in the homepage, the
`README.md` content.